blobcached: default files
